createTag and getRecentReleaseMovies
reworked createTag to include usernames, privacy, and movieIds

edited endpoint for createTags

added endpoint for getRecentReleaseMovies and editted the DatabaseController method to take the number to return as a parameter and the return a list of movies not documents.

This should fix Issue #84 since the tagNames got a serious rework. If this still does not resolve the problem we will keep hunting.