getTagScoresForMovieModal
added getTagScoresForMovieModal and endpoint meant to be hit when loading the movie modal to populate the tag portion. it returns a list of unique tags (hopefully in descending order) with the fields tagName, totalRating (a sum of the upvotes vs downvotes), and state (which is set to whatever the requester has up or downvoted each of the tags).