added getUniqueRatingCategoriesAndUserRatingWithMovieId
added endpoint used to populate the ratings portion of the rate movie modal. returns a list of Rating objects representing all the unique rating categories (combos of ratingName and upperbound) attached to a movie. If the requester has already rated the movie in one of the categories their rating is included as well in the userRating field.